+++ apps/assets/src/components/asset/dropdown/DropdownChains.tsx
@@ -89,7 +89,7 @@ const DropdownChains = ({
                         alt={option.chainIdentifier}
                         width={25}
                         height={25}
-                        className=" h-6 w-6"
+                        className="h-6 w-6 rounded-full"
                       />
                       <span>{option.chainIdentifier}</span>
                     </div>