+++ apps/assets/src/components/asset/dropdown/DropdownTokens.tsx
@@ -84,7 +84,7 @@ const DropdownTokens = ({
                       alt={option.symbol}
                       width={25}
                       height={25}
-                      className=" h-6 w-6"
+                      className="h-6 w-6 rounded-full"
                     />
                     <span>{option.symbol}</span>
                   </div>