+++ apps/assets/src/components/asset/table/ContentTable.tsx
@@ -51,8 +51,10 @@ const ContentTable = ({
     const map = new Map<string, accordionData>();
     tableData?.table.map((e) => {
       if (
-        e.chainIdentifier === "Stride" &&
-        map.has(e.chainIdentifier) === true
+        (e.chainIdentifier === "Stride" &&
+          map.has(e.chainIdentifier) === true) ||
+        (e.chainIdentifier === "Quicksilver" &&
+          map.has(e.chainIdentifier) === true)
       ) {
         const temp = map.get(e.chainIdentifier);
         if (temp === undefined) {
@@ -60,7 +62,10 @@ const ContentTable = ({
         }
         temp.tokens.push(e);
         temp.total = temp.total.add(e.erc20Balance);
-      } else if (e.chainIdentifier === "Stride") {
+      } else if (
+        e.chainIdentifier === "Stride" ||
+        e.chainIdentifier === "Quicksilver"
+      ) {
         map.set(e.chainIdentifier, {
           name: e.chainIdentifier,
           icon: e.chainIdentifier,