+++ apps/assets/src/components/asset/table/components/Description.tsx
@@ -25,7 +25,7 @@ export const Description = ({
         alt={symbol}
         width={30}
         height={30}
-        className=""
+        className="rounded-full"
       />
       <div className="flex flex-col items-start">
         <span className="font-bold">{symbol}</span>